requirements.txt
flask
pandas
gunicorn
sklearn
scikit-image
numpy
matplotlib.pyplot
plotly
scipy.sparse
seaborn

